python-bot/requirements.txt
python-dotenv
pyTelegramBotAPI
requests
python-telegram-bot


